backend/requirements.txt
django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
requests==2.33.1
psycopg[binary]==3.3.3
